ui: Statements table tooltips
Move tooltips contents in separate file

Release note (ui): none
