+++ pkg/ui/src/views/shared/components/sortabletable/sortabletable.styl
@@ -117,8 +117,7 @@
   color $colors--neutral-1
   white-space pre-wrap
   margin-bottom 0
-  line-height 22px
-  color $colors--neutral-6
+  cursor pointer
   span
     margin-right 6px
   a
