+++ pkg/ui/src/views/statements/statementsTable.tsx
@@ -8,24 +8,20 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-import getHighlightedText from "src/util/highlightedText";
 import React from "react";
-import { Link } from "react-router-dom";
 import classNames from "classnames/bind";
 
 import { StatementStatistics } from "src/util/appStats";
 import { FixLong } from "src/util/fixLong";
-import { StatementSummary, summarize } from "src/util/sql/summarize";
+import { StatementSummary } from "src/util/sql/summarize";
 import { ColumnDescriptor, SortedTable } from "src/views/shared/components/sortedtable";
 import { countBarChart, latencyBarChart, retryBarChart, rowsBarChart } from "./barCharts";
-import { Anchor, Tooltip } from "src/components";
 import "./statements.styl";
-import { DiagnosticStatusBadge } from "./diagnostics/diagnosticStatusBadge";
 import { cockroach } from "src/js/protos";
 import IStatementDiagnosticsReport = cockroach.server.serverpb.IStatementDiagnosticsReport;
 import { ActivateDiagnosticsModalRef } from "./diagnostics/activateDiagnosticsModal";
 import styles from "./statementsTable.module.styl";
-import { statementsSql, transactionalPipelining, statementDiagnostics, statementsRetries, statementsTimeInterval } from "oss/src/util/docs";
+import { StatementTableTitle, StatementTableCell, NodeNames } from "./statementsTableContent";
 
 const cx = classNames.bind(styles);
 const longToInt = (d: number | Long) => FixLong(d).toInt();
@@ -41,28 +37,6 @@ export interface AggregateStatistics {
 }
 
 export class StatementsSortedTable extends SortedTable<AggregateStatistics> {}
-
-function StatementLink(props: { statement: string, app: string, implicitTxn: boolean, search: string }) {
-  const summary = summarize(props.statement);
-  const base = props.app && props.app.length > 0 ? `/statements/${props.app}/${props.implicitTxn}` : `/statement/${props.implicitTxn}`;
-  return (
-    <Link to={ `${base}/${encodeURIComponent(props.statement)}` }>
-      <div className="cl-table-link__tooltip">
-        <Tooltip
-          placement="bottom"
-          title={<pre className="cl-table-link__description">
-            { getHighlightedText(props.statement, props.search) }
-          </pre>}
-          overlayClassName="cl-table-link__statement-tooltip--fixed-width"
-        >
-          <div className="cl-table-link__tooltip-hover-area">
-            { getHighlightedText(shortStatement(summary, props.statement), props.search, true) }
-          </div>
-        </Tooltip>
-      </div>
-    </Link>
-  );
-}
 
 export function shortStatement(summary: StatementSummary, original: string) {
   switch (summary.statement) {
@@ -84,69 +58,13 @@ export function makeStatementsColumns(
 ): ColumnDescriptor<AggregateStatistics>[]  {
   const columns: ColumnDescriptor<AggregateStatistics>[] = [
     {
-      title: (
-        <Tooltip
-          placement="bottom"
-          title={
-            <div className="tooltip__table--title">
-              <p>
-                {"SQL statement "}
-                <Anchor
-                  href={statementsSql}
-                  target="_blank"
-                >
-                  fingerprint.
-                </Anchor>
-              </p>
-              <p>
-                To view additional details of a SQL statement fingerprint, click this to open the Statement Details page.
-              </p>
-            </div>
-          }
-        >
-          Statements
-        </Tooltip>
-      ),
+      title: StatementTableTitle.statements,
       className: "cl-table__col-query-text",
-      cell: (stmt) => (
-        <StatementLink
-          statement={ stmt.label }
-          implicitTxn={ stmt.implicitTxn }
-          search={search}
-          app={ selectedApp }
-        />
-      ),
+      cell: StatementTableCell.statements(search, selectedApp),
       sort: (stmt) => stmt.label,
     },
     {
-      title: (
-        <Tooltip
-          placement="bottom"
-          title={
-            <div className="tooltip__table--title">
-              <p>
-                {"Type of transaction (implicit or explicit). Explicit transactions refer to statements that are wrapped by "}
-                <code>BEGIN</code>
-                {" and "}
-                <code>COMMIT</code>
-                {" statements by the client. Explicit transactions employ "}
-                <Anchor
-                  href={transactionalPipelining}
-                  target="_blank"
-                >
-                  transactional pipelining
-                </Anchor>
-                {" and therefore report latencies that do not account for replication."}
-              </p>
-              <p>
-                For statements not in explicit transactions, CockroachDB wraps each statement in individual implicit transactions.
-              </p>
-            </div>
-          }
-        >
-          TXN Type
-        </Tooltip>
-      ),
+      title: StatementTableTitle.txtType,
       className: "statements-table__col-time",
       cell: (stmt) => (stmt.implicitTxn ? "Implicit" : "Explicit"),
       sort: (stmt) => (stmt.implicitTxn ? "Implicit" : "Explicit"),
@@ -156,40 +74,8 @@ export function makeStatementsColumns(
 
   if (activateDiagnosticsRef) {
     const diagnosticsColumn: ColumnDescriptor<AggregateStatistics> = {
-      title: (
-        <Tooltip
-          placement="bottom"
-          title={
-            <div className="tooltip__table--title">
-              <p>
-                {"Option to activate "}
-                <Anchor
-                  href={statementDiagnostics}
-                  target="_blank"
-                >
-                  diagnostics
-                </Anchor>
-                {" for each statement. If activated, this displays the status of diagnostics collection ("}
-                <code>WAITING FOR QUERY</code>, <code>READY</code>, OR <code>ERROR</code>).
-              </p>
-            </div>
-          }
-        >
-          Diagnostics
-        </Tooltip>
-      ),
-      cell: (stmt) => {
-        if (stmt.diagnosticsReport) {
-          return <DiagnosticStatusBadge status={stmt.diagnosticsReport.completed ? "READY" : "WAITING FOR QUERY"}/>;
-        }
-        return (
-          <Anchor
-            onClick={() => activateDiagnosticsRef?.current?.showModalFor(stmt.label)}
-          >
-            Activate
-          </Anchor>
-        );
-      },
+      title: StatementTableTitle.diagnostics,
+      cell: StatementTableCell.diagnostics(activateDiagnosticsRef),
       sort: (stmt) => {
         if (stmt.diagnosticsReport) {
           return stmt.diagnosticsReport.completed ? "READY" : "WAITING FOR QUERY";
@@ -202,20 +88,12 @@ export function makeStatementsColumns(
   return columns;
 }
 
-function NodeLink(props: { nodeId: string, nodeNames: { [nodeId: string]: string } }) {
-  return (
-    <Link to={ `/node/${props.nodeId}` }>
-      <div className="node-name-tooltip__info-icon">{props.nodeNames[props.nodeId]}</div>
-    </Link>
-  );
-}
-
-export function makeNodesColumns(statements: AggregateStatistics[], nodeNames: { [nodeId: string]: string })
+export function makeNodesColumns(statements: AggregateStatistics[], nodeNames: NodeNames)
     : ColumnDescriptor<AggregateStatistics>[] {
   const original: ColumnDescriptor<AggregateStatistics>[] = [
     {
       title: null,
-      cell: (stmt) => <NodeLink nodeId={stmt.label} nodeNames={ nodeNames } />,
+      cell: StatementTableCell.nodeLink(nodeNames),
       // sort: (stmt) => stmt.label,
     },
   ];
@@ -263,112 +141,25 @@ function makeCommonColumns(statements: AggregateStatistics[])
 
   return [
     {
-      title: (
-        <Tooltip
-          placement="bottom"
-          title={
-            <div className="tooltip__table--title">
-              <p>
-                {"Cumulative number of "}
-                <Anchor
-                  href={statementsRetries}
-                  target="_blank"
-                >
-                  retries
-                </Anchor>
-                {" of statements with this fingerprint within the last hour or specified time interval."}
-              </p>
-            </div>
-          }
-        >
-          Retries
-        </Tooltip>
-      ),
+      title: StatementTableTitle.retries,
       className: "statements-table__col-retries",
       cell: retryBar,
       sort: (stmt) => (longToInt(stmt.stats.count) - longToInt(stmt.stats.first_attempt_count)),
     },
     {
-      title: (
-        <Tooltip
-          placement="bottom"
-          title={
-            <div className="tooltip__table--title">
-              <p>
-                {"Cumulative number of executions of statements with this fingerprint within the last hour or specified "}
-                <Anchor
-                  href={statementsTimeInterval}
-                  target="_blank"
-                >
-                  time interval
-                </Anchor>.
-              </p>
-              <p>
-                {"The bar indicates the ratio of runtime success (gray) to "}
-                <Anchor
-                  href={statementsRetries}
-                  target="_blank"
-                >
-                  retries
-                </Anchor>
-                {" (red) for the SQL statement fingerprint."}
-              </p>
-            </div>
-          }
-        >
-          Execution Count
-        </Tooltip>
-      ),
+      title: StatementTableTitle.executionCount,
       className: "statements-table__col-count",
       cell: countBar,
       sort: (stmt) => FixLong(stmt.stats.count).toInt(),
     },
     {
-      title: (
-        <Tooltip
-          placement="bottom"
-          title={
-            <div className="tooltip__table--title">
-              <p>
-                {"Average number of rows returned while executing statements with this fingerprint within the last hour or specified "}
-                <Anchor
-                  href={statementsTimeInterval}
-                  target="_blank"
-                >
-                  time interval
-                </Anchor>.
-              </p>
-              <p>
-                The gray bar indicates the mean number of rows returned. The blue bar indicates one standard deviation from the mean.
-              </p>
-            </div>
-          }
-        >
-          Rows Affected
-        </Tooltip>
-      ),
+      title: StatementTableTitle.rowsAffected,
       className: "statements-table__col-rows",
       cell: rowsBar,
       sort: (stmt) => stmt.stats.num_rows.mean,
     },
     {
-      title: (
-        <Tooltip
-          placement="bottom"
-          title={
-            <div className="tooltip__table--title">
-              <p>
-                Average service latency of statements with this fingerprint within the last hour or specified time interval.
-              </p>
-              <p>
-                The gray bar indicates the mean latency. The blue bar indicates one standard deviation from the mean.
-              </p>
-            </div>
-          }
-        >
-          Latency
-        </Tooltip>
-      ),
+      title: StatementTableTitle.latency,
       className: "statements-table__col-latency",
       cell: latencyBar,
       sort: (stmt) => stmt.stats.service_lat.mean,
